requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.10.0
django-on-heroku==1.1.2
django-rest-framework==0.1.0
djangorestframework==3.12.4
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dotenv==0.15.0
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.2.0
